This is the documentation file of VecG_TNR

The following is the function and documentations

Base.:* — Method

Overload the multiplication * to group multiplication.

Example

julia> GroupElement(2, CyclicGroup(3)) * GroupElement(1, CyclicGroup(3))
GroupElement{CyclicGroup}(0, CyclicGroup(3))

Base.:+ — Method

Elementwise addition of two morphisms.

Input:

two morphisms

Output:

the sum of the two morphisms

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T1 = random_mor(Float64, (A, B, C, D))
       T2 = random_mor(Float64, (A, B, C, D))
       T1 + T2

Base.:- — Method

Elementwise subtraction of two morphisms.

Input:

two morphisms

Output:

the difference of the two morphisms

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T1 = random_mor(Float64, (A, B, C, D))
       T2 = random_mor(Float64, (A, B, C, D))
       T1 - T2

Base.:== — Method

Overload the == to group case.

Base.:== — Method

Comparing two objects.

Base.Broadcast.broadcasted — Method

Elementwise division of a morphism by a number.

Input:

a morphism

a number

Output:

a new morphism created by performing division

Base.Broadcast.broadcasted — Method

Take the elementwise square root of a morphism.

Input:

a morphism

Output:

a new morphism created by taking the elementwise square root

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = random_mor(Float64, (A, B, C, D))
       sqrt(T)

Base.eltype — Method

Get the element type of the group.

Base.getindex — Method

Get the object of the i-th leg

Input:

a morphism

the index

Output:

the object of the i-th leg

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = Mor(Float64, (A, B, C, D))
       T[1]
       e ⊕ 2a ⊕ 3a²

Base.getindex — Method

Get a vector of objects in a morphism

Input:

a morphism

a range

Output:

a vector of objects in the morphism

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = Mor(Float64, (A, B, C, D))
       T[1:2]
       (e ⊕ 2a ⊕ 3a², e ⊕ 3a ⊕ 2a²)

Base.getindex — Method

Get the tensor of the sector

Input:

a morphism

a sector

Output:

the tensor of the sector

Both the form T[S] and T[S.sect], which can be, for example T[g,h,k,l], are supported.

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = Mor(Float64, (A, B, C, D))
       S = Sector(e, a, aa)
       T[S]
       T[e,a,aa]

Base.getindex — Method

Get the key-th group element in a sector

Input:

a sector

the index

Output:

the key-th group element in the sector

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       S = Sector(e, a, aa)
       S[2]
       a

Base.getindex — Method

Get the multiplicity of a group element in an object

Input:

an object

a group element

Output:

the multiplicity of the group element in the object

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       A[a]
       2

Base.lastindex — Method

Get the number of legs in a morphism

Input:

a morphism

Output:

the number of legs in the morphism

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = Mor(Float64, (A, B, C, D))
        lastindex(T)
       4

Base.setindex! — Method

Set the multiplicity of a group element in an object

Input:

an object

the multiplicity

a group element

Output:

the object with the multiplicity of the group element set

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       A[a] = 5
       e⊕5a⊕3a²

Base.setindex! — Method

Set the morphism to a given array for a given sector

Input:

a morphism

an array

a tuple of group elements or a sector

Output:

the morphism with the tensor set to the array for the sector constructed by the tuple of group elements

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = random_mor(Float64, (A, B, C, D))
       T[e,a,aa,e] = rand(1,3,3,2)
julia> S = Sector(e, a, aa, e)
       T[S] = rand(1,3,3,2)

Main.VecG_TNR.VecG_dag — Method

Take the dagger of a morphism. The dagger of a morphism is defined by taking the conjugate transpose of the tensor of each sector. The new legs are in the reverse order of the original legs. Moreover, the objects of the new morphism are the dual objects of the original morphism. For example:

The 1st leg becomes the last leg

The 2nd leg becomes the 2nd last leg

The 3rd leg becomes the 3rd last leg

...

Here is an picture to illustrate the idea:

| | | | |
1 2 3 4 5
| | | | |
^ ^ ^ ^ ^
| | | | |
    T

After taking the dagger, the morphism becomes:

| | | | |
5 4 3 2 1
| | | | |
v v v v v
| | | | |
    T^†

Input:

a morphism

Output:

a new morphism created by taking the dagger

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = random_mor(Float64, (A, B, C, D))
       dagger(T)

Main.VecG_TNR.VecG_permutedims — Method

Permute the dimensions of a morphism.

Input:

a morphism

a tuple of integers, which is a permutation

Output:

a morphism, whose dimensions are permuted

Example

julia> D6 = DihedralGroup(3)
    e = identity_element(D6)
    s = GroupElement((1,0), D6)
    r = GroupElement((0,1), D6)
    A = Obj(e=>2, s=>3, r=>2, s*r=>1)
    B = Obj(e=>2, s=>3, r=>2, s*r=>1)
    C = Obj(e=>2, s=>3, r=>2, s*r=>1)
    D = Obj(e=>2, s=>3, r=>2, s*r=>1)
    T = random_mor(Float64, (A, B, C, D))
    Tperm = VecG_permutedims(T, (2,3,4,1)) 
    Tperm[e,s,r,s*r] == permutedims(T[s*r, e, s, r],(2,3,4,1))
    true

Main.VecG_TNR.VecG_permutesectors — Method

Permute the sectors of a morphism.

Input:

a morphism

a tuple of integers, which is a permutation

Output:

a morphism, whose sectors are permuted

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       S = Sector(e, a, aa)
       VecG_permutesectors(S, (2,3,1))
        a ⊗ a² ⊗ e

Main.VecG_TNR.block_matrix_svd — Method

Perform SVD on a block matrix, where each block is a submatrix. The input is a 2D array of matrices, and the output is a tuple of block matrices U, S, and V. The block matrices U and V are the left and right singular vectors, respectively. The block matrix S is the singular values, stored as a vector.

Arguments

matrices::AbstractMatrix{<:AbstractMatrix}: A 2D array of matrices to be factorized.

Returns

U_blocks::Vector{AbstractMatrix}: A vector of block matrices representing the left singular vectors.

S::Vector: A vector of singular values.

V_blocks::Vector{AbstractMatrix}: A vector of block matrices representing the right singular vectors.

Example

A = [1 2; 3 4]  # 2x3 matrix
B = [5 6 6; 7 8 8]  # 2x3 matrix
C = [9 10; 11 12; 15 16; 17 18]  # 4x2 matrix
D = [13 14 19; 15 16 20; 21 22 23; 25 26 27]  # 4x3 matrix
matrices = Array{Matrix{Float64}}(undef, 2, 2)
matrices[1, 1] = A
matrices[1, 2] = B
matrices[2, 1] = C
matrices[2, 2] = D
U_blocks, S, V_blocks = block_matrix_svd(matrices)

The output Ublocks, S, and Vblocks can be compared with the output of the svd function. Here we can use the concatenatematriceswithmetadata function to concatenate the block matrices into a single large matrix and compare the results. However, before doing so, we need to reshape the block matrices into a 2x1 array to match the input format of the concatenatematriceswithmetadata function. The comparison should return true for both U and V.

Ucat, _ = concatenate_matrices_with_metadata(AbstractMatrix{AbstractMatrix}(reshape(U_blocks,2,1)))
Vcat, _ = concatenate_matrices_with_metadata(AbstractMatrix{AbstractMatrix}(reshape(V_blocks,2,1)))
U, s, V = svd([A B; C D])
@show Ucat ≈ U
@show Vcat ≈ V
true
true

Thus we would like to expect that the Ublocks * S * Vblocks' is approximately equal to the original matrix. The Hermitian conjugate need to be taken into account when dealing with complex matrices. This choice of the convension if different from the ITensor package, where the Hermitian conjugate is taken into account in the SVD function. In VecG tensor svd, we will include the Hermitian conjugate in the SVD function, such that the contraction of U * S * V will directly give the original matrix.

Main.VecG_TNR.concatenate_matrices_with_metadata — Method

Concatenate a 2D array of matrices into a single large matrix, and record metadata for the submatrices. The metadata is stored in a dictionary with keys (i, j), which represents the location of the block. And values (startrow, startcol, numrows, numcols) represent the starting row and column of the block, and the number of rows and columns of the block. The metadata can be used to extract the submatrices from the large matrix.

Arguments

matrices::AbstractMatrix{<:AbstractMatrix}: A 2D array of matrices to be concatenated.

Returns

big_matrix::Matrix: The large matrix obtained by concatenating the input matrices.

metadata::Dict{Tuple{Int, Int}, Tuple{Int, Int, Int, Int}}: The metadata for the submatrices.

Example

M = Matrix{Matrix}(undef, 2, 2)
A = [1 2; 3 4]  # 2x2 matrix
B = [5 6 6; 7 8 8]  # 2x3 matrix
C = [9 10; 11 12; 15 16; 17 18]  # 3x2 matrix
D = [13 14 19; 15 16 20; 21 22 23]  # 3x3 matrix
M[1,1] = A
M[1,2] = B
M[2,1] = C
M[2,2] = D
big_matrix, metadata = concatenate_matrices_with_metadata(M)
@show big_matrix
6×5 Matrix{Int64}:
  1   2   5   6   6
  3   4   7   8   8
  9  10  13  14  19
 11  12  15  16  20
 15  16  21  22  23
 17  18   0   0   0
 @show metadata
 Dict{Tuple{Int64, Int64}, NTuple{4, Int64}} with 4 entries:
  (1, 2) => (1, 3, 2, 3)
  (1, 1) => (1, 1, 2, 2)
  (2, 2) => (3, 3, 3, 3)
  (2, 1) => (3, 1, 4, 2)

Main.VecG_TNR.dual_obj — Method

Generating a dual object.

Input:

an object

Output:

the dual object, whose multiplicities are the same as the input object, but the group elements are inversed.

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       dual_obj(A)

e⊕3a⊕2a²

Main.VecG_TNR.elements — Method

Generating the tuple of elements of group. Support CyclicGroup, DihedralGroup and ProductGroup. However, for integer group, the elements are infinite, so we do not provide this function.

Example

julia> elements(CyclicGroup(3))
(GroupElement{CyclicGroup}(0,CyclicGroup(3)), GroupElement{CyclicGroup}(1,CyclicGroup(3)), GroupElement{CyclicGroup}(2,CyclicGroup(3)))

julia> elements(DihedralGroup(3))
(GroupElement{DihedralGroup}((0,0), DihedralGroup(3)), GroupElement{DihedralGroup}((0,1), DihedralGroup(3)), GroupElement{DihedralGroup}((0,2), DihedralGroup(3)), GroupElement{DihedralGroup}((1,0), DihedralGroup(3)), GroupElement{DihedralGroup}((1,1), DihedralGroup(3)), GroupElement{DihedralGroup}((1,2), DihedralGroup(3)))

julia> elements(ProductGroup(CyclicGroup(2), CyclicGroup(2)))
((GroupElement{CyclicGroup}(0, CyclicGroup(2)), GroupElement{CyclicGroup}(1, CyclicGroup(2))), (GroupElement{CyclicGroup}(0, CyclicGroup(2)), GroupElement{CyclicGroup}(1, CyclicGroup(2))))

Main.VecG_TNR.extract_blocks_to_matrix — Method

Extracts VecG morphism to block matrices when given a fixed group element g_bridge. This is an intermediate function before performing svd.

Input

A morphism

A integer, before and include which legs will be considered as row indices, and after which legs will be considered as column indices

A dictionary that representing the splitting of the sector, produced from tosectoroutin

Output:

A dictionary, whose keys are bridge sectors appeared in the morphism, and values are matrix of matrices

Example

julia> D6 = DihedralGroup(3)
        e = identity_element(D6)
        s = GroupElement((1,0), D6)
        r = GroupElement((0,1), D6)
        A = Obj(e=>2, s=>3, r=>2, s*r=>1) 
        B = Obj(e=>2, s*r*r=>2, r=>3, s*r=>2)
        C = Obj(e=>2, s=>3, r*r=>2, s*r=>2)
        D = Obj(e=>2, s=>4, r=>1, s*r*r=>1)
        TT = random_mor(Float64, (A, B, C, D))
        sec_split = to_sector_outin(TT, 2)
        Mat = extract_blocks_to_matrix(TT, 2, sec_split)
        @show Mat[s][1,1] == reshape(TT[s,e,r*r,s*r*r], 6,2) # true
        @show Mat[s][1,2] == reshape(TT[s,e,s,e], 6,6) # true
        @show Mat[s][2,2] == reshape(TT[r,s*r,s,e], 4,6) # true

Main.VecG_TNR.get_group — Method

Get the group of a morphism

Input:

a morphism

Output:

the group of the morphism

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = Mor(Float64, (A, B, C, D))
       get_group(T)
       ℤ₃

Main.VecG_TNR.get_group — Method

Get the group that the object belongs to

Input:

an object

Output:

the group that the object belongs to

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       get_group(A)

Main.VecG_TNR.get_sector_size — Method

Get the size of the tensor for a given sector

Input:

a morphism

a tuple of group elements or a sector

Output:

the size of the tensor for the sector constructed by the tuple of group elements

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = random_mor(Float64, (A, B, C, D))
       get_sector_size(T, (e, a, aa, e))
       (1, 3, 3, 2)

julia> S = Sector(e, a, aa, e)
    get_sector_size(T, S)
    (1, 3, 3, 2)

Main.VecG_TNR.group_iter — Method

Generating a iterator, whose terms are n-tuple of group elements. It supports CyclicGroup, DihedralGroup and ProductGroup. However, since the elements of IntegerGroup are infinite, we do not provide this function.

Example

julia>  for i in group_iter(CyclicGroup(2), 2)
            println(i)
        end
(GroupElement{CyclicGroup}(0, CyclicGroup(2)), GroupElement{CyclicGroup}(0, CyclicGroup(2)))
(GroupElement{CyclicGroup}(1, CyclicGroup(2)), GroupElement{CyclicGroup}(0, CyclicGroup(2)))
(GroupElement{CyclicGroup}(0, CyclicGroup(2)), GroupElement{CyclicGroup}(1, CyclicGroup(2)))
(GroupElement{CyclicGroup}(1, CyclicGroup(2)), GroupElement{CyclicGroup}(1, CyclicGroup(2)))

Main.VecG_TNR.group_multiply — Method

group_multiply(x::GroupElement{CyclicGroup}, y::GroupElement{CyclicGroup})

Multiplying two group elements of cyclic group.

Example

julia> group_multiply(GroupElement(2, CyclicGroup(3)), GroupElement(1, CyclicGroup(3)))
GroupElement{CyclicGroup}(0, CyclicGroup(3))

Main.VecG_TNR.group_multiply — Method

group_multiply(x::GroupElement{DihedralGroup}, y::GroupElement{DihedralGroup})

Multiplying two group elements of dihedral group.

Example

julia> group_multiply(GroupElement((0,1), DihedralGroup(3)), GroupElement((1,0), DihedralGroup(3)))
GroupElement{DihedralGroup}((1,2), DihedralGroup(3))

Main.VecG_TNR.group_multiply — Method

group_multiply(x::GroupElement{IntegerGroup}, y::GroupElement{IntegerGroup})

Multiplying two group elements of integer group.

Example

julia> group_multiply(GroupElement(2, IntegerGroup()), GroupElement(3, IntegerGroup()))
GroupElement{IntegerGroup}(5, IntegerGroup())

Main.VecG_TNR.group_multiply — Method

group_multiply(x::GroupElement{ProductGroup}, y::GroupElement{ProductGroup})

Multiplying two group elements of product group.

Example

julia> group_multiply(GroupElement((GroupElement(1, CyclicGroup(3)), GroupElement((0,1), DihedralGroup(3))), ProductGroup(CyclicGroup(3), DihedralGroup(3))), GroupElement((GroupElement(2, CyclicGroup(3)), GroupElement((1,2), DihedralGroup(3))), ProductGroup(CyclicGroup(3), DihedralGroup(3)))
GroupElement{ProductGroup}((GroupElement{CyclicGroup}(0, CyclicGroup(3)), GroupElement{DihedralGroup}((1,0), DihedralGroup(3))), ProductGroup((CyclicGroup(3), DihedralGroup(3)))

Main.VecG_TNR.group_tree — Method

Generating an iterator of n-tuple of group elements, such that all elements are multiplied to g. It supports CyclicGroup, DihedralGroup and ProductGroup. However, since the elements of IntegerGroup are infinite, we do not provide this function.

Input Parameter:

g: a group element

n: the number of group elements in a tuple

Example

julia> x = group_tree(GroupElement(0, CyclicGroup(2)), 2)

Base.Generator{Base.Iterators.ProductIterator{Tuple{Tuple{GroupElement{CyclicGroup}, GroupElement{CyclicGroup}}}}, var"#18#20"{GroupElement{CyclicGroup}}}(var"#18#20"{GroupElement{CyclicGroup}}(GroupElement{CyclicGroup}(0, CyclicGroup(2))), Base.Iterators.ProductIterator{Tuple{Tuple{GroupElement{CyclicGroup}, GroupElement{CyclicGroup}}}}(((GroupElement{CyclicGroup}(0, CyclicGroup(2)), GroupElement{CyclicGroup}(1, CyclicGroup(2))),)))

julia> for i in x
            println(i)
       end

(GroupElement{CyclicGroup}(0, CyclicGroup(2)), GroupElement{CyclicGroup}(0, CyclicGroup(2)))
(GroupElement{CyclicGroup}(1, CyclicGroup(2)), GroupElement{CyclicGroup}(1, CyclicGroup(2)))

Main.VecG_TNR.identity_element — Method

Give the identity element of the group

Example

julia> identity_element(CyclicGroup(3))
GroupElement{CyclicGroup}(0, CyclicGroup(3))

julia> identity_element(DihedralGroup(3))
GroupElement{DihedralGroup}((0,0),DihedralGroup(3))

julia> identity_element(ProductGroup(CyclicGroup(3), DihedralGroup(3)))
GroupElement{ProductGroup}((GroupElement{CyclicGroup}(0, CyclicGroup(3)), GroupElement{DihedralGroup}((0,0), DihedralGroup(3))), ProductGroup((CyclicGroup(3), DihedralGroup(3)))

julia> identity_element(IntegerGroup())
GroupElement{IntegerGroup}(0, IntegerGroup())

Main.VecG_TNR.identity_mor — Method

Generating an identity morphism for a given object.

Input:

element type: data type, for example Float64, ComplexF64, etc...

object: an object

Output:

an identity morphism, whose data are identity matrices

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       identity_mor(Float64, A)

Main.VecG_TNR.inverse — Method

Give the inverse of the element x

Example

julia> inverse(GroupElement(1, CyclicGroup(3)))
GroupElement{CyclicGroup}(2, CyclicGroup(3))

julia> inverse(GroupElement((0,2), DihedralGroup(3)))
GroupElement{DihedralGroup}((0,1), DihedralGroup(3))

julia> inverse(GroupElement((GroupElement(1, CyclicGroup(3)), GroupElement((0,2), DihedralGroup(3))), ProductGroup(CyclicGroup(3), DihedralGroup(3)))
GroupElement{ProductGroup}((GroupElement{CyclicGroup}(2, CyclicGroup(3)), GroupElement{DihedralGroup}((0,1), DihedralGroup(3))), ProductGroup((CyclicGroup(3), DihedralGroup(3)))

julia> inverse(GroupElement(1, IntegerGroup()))
GroupElement{IntegerGroup}(-1, IntegerGroup())

Main.VecG_TNR.is_accend — Method

Judge whether a tuple is in an accending order.

Input:

a tuple of integers

a modulus

Output:

a boolean value, indicating whether the tuple is in an accending order

It reports false if the tuple is not in an accending order or the elements are not in the range of 1 to modn.

Example

julia> is_accend((1,2,3,4), 4)
       true

julia> is_accend((4,1), 4)
    true

Main.VecG_TNR.is_cyclic — Method

Judge whether a tuple is cyclic.

Input:

a tuple of integers

a modulus

Output:

a boolean value, indicating whether the tuple is cyclic

It reports false if the tuple is not cyclic or the elements are not in the range of 1 to modn.

Example

julia> is_cyclic((1,2,3,4), 4)
       true

julia> is_cyclic((4,1), 4)
         false

Main.VecG_TNR.is_descend — Method

Judge whether a tuple is in a descending order.

Input:

a tuple of integers

a modulus

Output:

a boolean value, indicating whether the tuple is in a descending order

It reports false if the tuple is not in a descending order or the elements are not in the range of 1 to modn.

Example

julia> is_descend((4,3,2,1), 4)
       true

julia> is_descend((4,1), 4) 
         false

Main.VecG_TNR.max_abs — Method

Find the maximum absolute value of the element in a morphism.

Input:

a morphism

Output:

the maximum absolute value of the element in the morphism

Example

julia> G = CyclicGroup(2)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       A = Obj(e=>1, a=>1)
       B = Obj(e=>1, a=>1)
         T = Mor(Float64, (A, B))
         T[e,e] = reshape([1],1,1)
            T[a,a] = reshape([2],1,1)
       max_abs(T)
         2.0

Main.VecG_TNR.multiply — Method

Multiplying a tuple of group elements.

Example

julia> multiply((GroupElement(1, CyclicGroup(4)), GroupElement(2, CyclicGroup(4))))
GroupElement{CyclicGroup}(3, CyclicGroup(4))

Main.VecG_TNR.random_mor — Method

Generating a random morphism for a given objects.

Input:

element type: data type, for example Float64, ComplexF64, etc...

objects: a tuple of objects

Output:

a random morphism, whose data are random numbers

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = random_mor(Float64, (A, B, C, D))
         The data is too long to show. 

Main.VecG_TNR.to_perm — Method

Convert a tuple to a permutation.

Input:

a tuple of integers (i,j,k,...)

a modulus n

Output:

a tuple of integers, which is a permutation of the input tuple (i,j,k,..., 1,2,..., i-1)

Example

julia> to_perm((1,2,3), 4)
       (1, 2, 3, 4)

julia> to_perm((3,4,1), 4)
       (3, 4, 1, 2)

Main.VecG_TNR.to_sector_outin — Method

Given a bridge group element gbridge, find all sectors that fuses to gbridge when splited to 1, ..., nlegsplit-1, nlegsplit, | nlegsplit+1, nlegsplit+2, ...

Graphically, we consider the following:

|  |  |  |  |
g1 g2...    g_n_leg_split
|  |  |  |  |
^  ^  ^  ^  ^
|  |  |  |  |
 \  \ | /  /
      ^
      |
      g_bridge
      |
 /  / | \  \
|  |  |  |  |
gn gn-1...g_n_leg_split+1  
|  |  |  |  |
v  v  v  v  v
|  |  |  |  |

Input:

a morphism

an integer that describes the number of splitted legs

Output:

a dictionary, whose keys are bridge sectors appeared in the morphism, and values are a 2-argument tuple, each of which is also a tuple, representing the out sectors and in sectors

Example

    D6 = DihedralGroup(3)
    e = identity_element(D6)
    s = GroupElement((1,0), D6)
    r = GroupElement((0,1), D6)
    A = Obj(e=>2, s=>3, r=>2, s*r=>15) 
    B = Obj(e=>2, s*r*r=>4, r=>3, s*r=>2)
    C = Obj(e=>2, s=>3, r*r=>2, s*r=>2)
    D = Obj(e=>2, s=>4, r=>3, s*r*r=>15)
    TT = random_mor(Float64, (A, B, C, D))
    to_sector_outin(TT, 2)
    Dict{Any, Any} with 6 entries:
    s   => (((s, e), (r, sr)), ((r², sr²), (s, e), (e, s)))
    e   => (((e, e), (sr, sr)), ((e, e), (r², r), (s, s)))
    sr  => (((r, sr²), (sr, e), (s, r), (e, sr)), ((sr, e), (s, r), (r², s)))
    r   => (((r, e), (sr, sr²), (e, r), (s, sr)), ((s, sr²), (r², e), (sr, s)))
    r²  => (((s, sr²), (r, r)), ((e, r), (sr, sr²)))
    sr² => (((sr, r), (e, sr²)), ((e, sr²), (sr, r)))

Main.VecG_TNR.verify_group_axioms — Method

Helper to verify group axioms.

Input:

g: group

Main.VecG_TNR.zero_mor — Method

Generating a zero morphism for a given objects.

Input:

element type: data type, for example Float64, ComplexF64, etc...

objects: a tuple of objects

Output:

a zero morphism, whose data are all zeros

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       A = Obj(e=>1, a=>2, aa=>3)
       B = Obj(e=>2, a=>3, aa=>2)
       C = Obj(e=>1, a=>2, aa=>3)
       D = Obj(e=>2, a=>3, aa=>2)
       T = zero_mor(Float64, (A, B, C, D))
         The data is too long to show.

Main.VecG_TNR.zero_obj — Method

Generating a zero object.

Input:

a group

Output:

an object with all zero multiplicities

Example

``` julia> G = CyclicGroup(3) zero_obj(G)

Main.VecG_TNR.CyclicGroup — Type

CyclicGroup(n)

Construct a structure of group ℤn.

Example

julia> CyclicGroup(4)
CyclicGroup(4)

Main.VecG_TNR.DihedralGroup — Type

DihedralGroup(n)

Construct a structure of group D2n.

Example

julia> DihedralGroup(4)
DihedralGroup(4)

Main.VecG_TNR.GroupElement — Type

Definition of the structure GroupElement

Example

julia> GroupElement(2, CyclicGroup(4))
GroupElement{CyclicGroup}(2, CyclicGroup(4))

julia> GroupElement((2,1), DihedralGroup(4))
GroupElement{DihedralGroup}((2,1), DihedralGroup(4))

Main.VecG_TNR.GroupElement — Method

Construct a group element of group, with value value.

Parameter:

value: label of group element

group: abstract group structure

Example

julia> GroupElement(2, CyclicGroup(4))
GroupElement{CyclicGroup}(2, CyclicGroup(4))

julia> GroupElement((2,1), DihedralGroup(4))
GroupElement{DihedralGroup}((2,1), DihedralGroup(4))

julia> GroupElement((GroupElement(2, CyclicGroup(4), GroupElement((2,1), DihedralGroup(4)))), ProductGroup(CyclicGroup(4), DihedralGroup(4)))
GroupElement{ProductGroup}((GroupElement{CyclicGroup}(2, CyclicGroup(4)), GroupElement{DihedralGroup}((2,1), DihedralGroup(4))), ProductGroup((CyclicGroup(4), DihedralGroup(4)))

julia> GroupElement(4, IntegerGroup())
GroupElement{IntegerGroup}(4, IntegerGroup())

Main.VecG_TNR.IntegerGroup — Type

Constructing integer group ℤ

julia> IntegerGroup()
IntegerGroup()

Main.VecG_TNR.Mor — Type

The key data type in VecG_TNR.

Components:

objects: a tuple of objects

data: a dictionary, whose keys are sectors and values are tensors

Construction function:

Mor(element_type, objects)

Input:

element type: data type, for example Float64, ComplexF64, etc...

objects: a tuple of objects

Output:

an empty tensor, stored by a dictionary, whose key are sectors

Main.VecG_TNR.Obj — Type

Structure of objects in VecG. To be used in Mor.

To adapt to infinite groups, we use a dictionary to store the multiplicities of group elements. Moreover, some group elements may not appear in the object.

Input:

pairs of the form g=>n

Output:

an object of the form n1g1⊕n2g2⊕n3g3...

Example

julia> G = CyclicGroup(3)
       e = GroupElement(0, G)
       a = GroupElement(1, G)
       aa = GroupElement(2, G)
       Obj(e=>1, a=>2, aa=>3)

e⊕2a⊕3a²

Main.VecG_TNR.ProductGroup — Type

ProductGroup(groups...)

Construct a structure of group G₁ × G₂ × ... × Gₙ.

Example

julia> ProductGroup(CyclicGroup(2), CyclicGroup(3))
ProductGroup((CyclicGroup(2), CyclicGroup(3)))

Main.VecG_TNR.Sector — Type

Structure of sectors in VecG

Input:

a tuple of group elements

Output:

a sector of the form g1⊗g2⊗g3..., such that g1g2g3... = e

If the sector is not consistent, an error will be thrown.

Example

julia> G = CyclicGroup(3)
       e = GroupElement((0,0), G)
       r = GroupElement((0,1), G)
       s = GroupElement((1,0), G)
       Sector(s, r, s*r, e)

s⊗r⊗sr⊗e